sandbox/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
asgiref==3.8.1
    # via django
asn1==3.1.0
    # via django-ldaporm
asttokens==3.0.0
    # via stack-data
crispy-bootstrap5==2025.6
    # via
    #   demo (pyproject.toml)
    #   django-wildewidgets
decorator==5.1.1
    # via ipython
django==5.1.4
    # via
    #   demo (pyproject.toml)
    #   crispy-bootstrap5
    #   django-appconf
    #   django-braces
    #   django-compressor
    #   django-crequest
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-wildewidgets
    #   django-xff
    #   djangorestframework
django-appconf==1.0.6
    # via django-compressor
django-braces==1.16.0
    # via demo (pyproject.toml)
django-compressor==4.5.1
    # via demo (pyproject.toml)
django-crequest==2018.5.11
    # via demo (pyproject.toml)
django-crispy-forms==2.4
    # via
    #   crispy-bootstrap5
    #   django-wildewidgets
django-debug-toolbar==4.4.6
    # via demo (pyproject.toml)
django-environ==0.11.2
    # via demo (pyproject.toml)
django-extensions==3.2.3
    # via demo (pyproject.toml)
django-filter==25.1
    # via demo (pyproject.toml)
./django-ldaporm.tar.gz
    # via demo (pyproject.toml)
django-sass-processor==1.4.1
    # via demo (pyproject.toml)
django-theme-academy==0.3.11
    # via demo (pyproject.toml)
django-wildewidgets==1.1.6
    # via
    #   demo (pyproject.toml)
    #   django-theme-academy
django-xff==1.3.0
    # via demo (pyproject.toml)
djangorestframework==3.16.0
    # via demo (pyproject.toml)
enum-compat==0.0.3
    # via asn1
executing==2.1.0
    # via stack-data
gunicorn==23.0.0
    # via demo (pyproject.toml)
ipython==8.30.0
    # via demo (pyproject.toml)
jedi==0.19.2
    # via ipython
ldap-filter==1.0.1
    # via django-ldaporm
libsass==0.23.0
    # via demo (pyproject.toml)
matplotlib-inline==0.1.7
    # via ipython
mysqlclient==2.2.6
    # via demo (pyproject.toml)
packaging==24.2
    # via gunicorn
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
prompt-toolkit==3.0.48
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   django-ldaporm
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.4.2
    # via python-ldap
pygments==2.18.0
    # via ipython
python-ldap==3.4.4
    # via django-ldaporm
pytz==2025.2
    # via django-ldaporm
rcssmin==1.1.2
    # via django-compressor
redis==5.2.1
    # via demo (pyproject.toml)
rjsmin==1.2.2
    # via django-compressor
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
structlog==24.2.0
    # via demo (pyproject.toml)
tabulate==0.9.0
    # via demo (pyproject.toml)
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via ipython
wcwidth==0.2.13
    # via prompt-toolkit
